+++ src/rpm/package.rs
@@ -1,5 +1,7 @@
 use std::{
-    fs, io,
+    fs,
+    io::{self, Read, Write},
+    os::unix::fs::PermissionsExt,
     path::{Path, PathBuf},
     str::FromStr,
 };
@@ -7,7 +9,7 @@ use std::{
 use digest::Digest;
 use num_traits::FromPrimitive;
 
-use crate::{constants::*, errors::*, CompressionType};
+use crate::{constants::*, decompress_stream, errors::*, CompressionType};
 
 #[cfg(feature = "signature-pgp")]
 use crate::signature::pgp::Verifier;
@@ -23,7 +25,7 @@ use super::Lead;
 ///
 /// Can either be created using the [`PackageBuilder`](crate::PackageBuilder)
 /// or used with [`parse`](`self::Package::parse`) to obtain from a file.
-#[derive(Debug)]
+#[derive(Clone, Debug, PartialEq)]
 pub struct Package {
     /// Header and metadata structures.
     ///
@@ -61,6 +63,94 @@ impl Package {
         self.write(&mut io::BufWriter::new(fs::File::create(path)?))
     }
 
+    /// Iterate over the file contents of the package payload
+    ///
+    /// # Examples
+    /// ```ignore
+    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
+    /// let package = rpm::Package::open("test_assets/freesrp-udev-0.3.0-1.25.x86_64.rpm")?;
+    /// for entry in package.files()? {
+    ///     let file = entry?;
+    ///     // do something with file.content
+    ///     println!("{} is {} bytes", file.metadata.path.display(), file.content.len());
+    /// }
+    /// # Ok(()) }
+    /// ```
+    pub fn files(&self) -> Result<FileIterator, Error> {
+        let file_entries = self.metadata.get_file_entries()?;
+        let archive = decompress_stream(
+            self.metadata.get_payload_compressor()?,
+            io::Cursor::new(self.content.clone()),
+        )?;
+
+        Ok(FileIterator {
+            file_entries,
+            archive,
+            count: 0,
+        })
+    }
+
+    /// Extract all contents of the package payload to a given directory
+    ///
+    /// # Examples
+    /// ```ignore
+    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
+    /// let package = rpm::Package::open("test_assets/ima_signed.rpm")?;
+    /// package.extract(&package.metadata.get_name()?)?;
+    /// # Ok(()) }
+    /// ```
+    pub fn extract(&self, dest: impl AsRef<Path>) -> Result<(), Error> {
+        fs::create_dir_all(&dest)?;
+
+        let dirs = self
+            .metadata
+            .header
+            .get_entry_data_as_string_array(IndexTag::RPMTAG_DIRNAMES)?;
+
+        // pull every base directory name in the package and create the directory in advance
+        for dir in dirs {
+            let dir_path = dest
+                .as_ref()
+                .join(Path::new(dir).strip_prefix("/").unwrap_or(dest.as_ref()));
+            fs::create_dir_all(&dir_path)?;
+        }
+
+        // TODO: reduce memory by replacing this with an impl that writes the files immediately after reading them from the archive
+        // instead of reading each file entirely into memory (while the archive is also entirely in memory) before writing them
+        for file in self.files()? {
+            let file = file?;
+            let file_path = dest.as_ref().join(
+                file.metadata
+                    .path
+                    .strip_prefix("/")
+                    .unwrap_or(dest.as_ref()),
+            );
+
+            let perms = fs::Permissions::from_mode(file.metadata.mode.permissions().into());
+            match file.metadata.mode {
+                FileMode::Dir { .. } => {
+                    fs::create_dir_all(&file_path)?;
+                    fs::set_permissions(&file_path, perms)?;
+                }
+                FileMode::Regular { .. } => {
+                    let mut f = fs::File::create(&file_path)?;
+                    f.write_all(&file.content)?;
+                    fs::set_permissions(&file_path, perms)?;
+                }
+                FileMode::SymbolicLink { .. } => {
+                    // broken symlinks (common for debuginfo handling) are perceived as not existing by "exists()"
+                    if file_path.exists() || file_path.symlink_metadata().is_ok() {
+                        fs::remove_file(&file_path)?;
+                    }
+                    std::os::unix::fs::symlink(&file.metadata.linkto, &file_path)?;
+                }
+                _ => unreachable!("Encountered an unknown or invalid FileMode"),
+            }
+        }
+
+        Ok(())
+    }
+
     /// Create package signatures using an external key and add them to the signature header
     #[cfg(feature = "signature-meta")]
     pub fn sign<S>(&mut self, signer: S) -> Result<(), Error>
@@ -79,8 +169,8 @@ impl Package {
     /// # Examples
     /// ```
     /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// let mut package = rpm::Package::open("test_assets/ima_signed.rpm")?;
-    /// let raw_secret_key = std::fs::read("./test_assets/secret_key.asc")?;
+    /// let mut package = rpm::Package::open("tests/assets/RPMS/noarch/rpm-basic-2.3.4-5.el9.noarch.rpm")?;
+    /// let raw_secret_key = std::fs::read("./tests/assets/signing_keys/secret_rsa4096.asc")?;
     /// let signer = rpm::signature::pgp::Signer::load_from_asc_bytes(&raw_secret_key)?;
     /// // It's recommended to use timestamp of last commit in your VCS
     /// let source_date = 1_600_000_000;
@@ -200,7 +290,6 @@ impl Package {
                 "signature_header(header and content)",
                 signature_header_and_content,
             );
-            use io::Read;
             let header_and_content_cursor =
                 io::Cursor::new(&header_bytes).chain(io::Cursor::new(&self.content));
             verifier.verify(header_and_content_cursor, signature_header_and_content)?;
@@ -298,7 +387,7 @@ impl Package {
     }
 }
 
-#[derive(PartialEq, Debug)]
+#[derive(Clone, Debug, PartialEq)]
 pub struct PackageMetadata {
     pub lead: Lead,
     pub signature: Header<IndexSignatureTag>,
@@ -671,7 +760,7 @@ impl PackageMetadata {
     ///
     /// ```
     /// # use rpm::Package;
-    /// # let package = Package::open("test_assets/389-ds-base-devel-1.3.8.4-15.el7.x86_64.rpm").unwrap();
+    /// # let package = Package::open("tests/assets/RPMS/noarch/rpm-basic-2.3.4-5.el9.noarch.rpm").unwrap();
     /// let offsets = package.metadata.get_package_segment_offsets();
     /// let lead = offsets.lead..offsets.signature_header;
     /// let sig_header = offsets.signature_header..offsets.header;
@@ -743,6 +832,16 @@ impl PackageMetadata {
     }
 
     /// Extract a the set of contained file names.
+    ///
+    /// # Examples
+    /// ```
+    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
+    /// let package = rpm::Package::open("test_assets/ima_signed.rpm")?;
+    /// for path in package.metadata.get_file_paths()? {
+    ///     println!("{}", path.display());
+    /// }
+    /// # Ok(()) }
+    /// ```
     pub fn get_file_paths(&self) -> Result<Vec<PathBuf>, Error> {
         // reconstruct the messy de-constructed paths
         let basenames = self
@@ -797,8 +896,7 @@ impl PackageMetadata {
 
     /// The digest algorithm used per file.
     ///
-    /// Note that this is not necessarily the same as the digest
-    /// used for headers.
+    /// Note that this is not necessarily the same as the digest used for headers.
     pub fn get_file_digest_algorithm(&self) -> Result<DigestAlgorithm, Error> {
         self.header
             .get_entry_data_as_u32(IndexTag::RPMTAG_FILEDIGESTALGO)
@@ -810,7 +908,17 @@ impl PackageMetadata {
             })
     }
 
-    /// Extract a the set of contained file names including the additional metadata.
+    /// Get a list of metadata about the files in the RPM, without the file contents.
+    ///
+    /// # Examples
+    /// ```
+    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
+    /// let package = rpm::Package::open("test_assets/ima_signed.rpm")?;
+    /// for entry in package.metadata.get_file_entries()? {
+    ///     println!("{} is {} bytes", entry.path.display(), entry.size);
+    /// }
+    /// # Ok(()) }
+    /// ```
     pub fn get_file_entries(&self) -> Result<Vec<FileEntry>, Error> {
         // rpm does not encode it, if it is the default md5
         let algorithm = self
@@ -863,6 +971,7 @@ impl PackageMetadata {
             Err(Error::TagNotFound(_)) => Ok(None),
             Err(e) => return Err(e),
         };
+        // TODO: verify this is correct behavior for links?
         let links = self
             .header
             .get_entry_data_as_string_array(IndexTag::RPMTAG_FILELINKTOS);
@@ -872,10 +981,7 @@ impl PackageMetadata {
         {
             Ok(ima_signatures) => Ok(Some(ima_signatures)),
             Err(Error::TagNotFound(_)) => Ok(None),
-            Err(e) => {
-                println!("{e:?}");
-                return Err(e);
-            }
+            Err(e) => return Err(e),
         };
 
         match (
@@ -1016,6 +1122,56 @@ impl PackageMetadata {
                 description?;
                 unreachable!()
             }
+        }
+    }
+}
+
+pub struct FileIterator<'a> {
+    file_entries: Vec<FileEntry>,
+    archive: Box<dyn io::Read + 'a>,
+    count: usize,
+}
+
+#[derive(Debug)]
+pub struct RpmFile {
+    pub metadata: FileEntry,
+    pub content: Vec<u8>,
+}
+
+impl Iterator for FileIterator<'_> {
+    type Item = Result<RpmFile, Error>;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.count >= self.file_entries.len() {
+            return None;
+        }
+
+        let file_entry = self.file_entries[self.count].clone();
+        self.count += 1;
+
+        let reader = cpio::NewcReader::new(&mut self.archive);
+
+        match reader {
+            Ok(mut entry_reader) => {
+                if entry_reader.entry().is_trailer() {
+                    return None;
+                }
+
+                let mut content = Vec::new();
+
+                if let Err(e) = entry_reader.read_to_end(&mut content) {
+                    return Some(Err(Error::Io(e)));
+                }
+                if let Err(e) = entry_reader.finish() {
+                    return Some(Err(Error::Io(e)));
+                }
+
+                Some(Ok(RpmFile {
+                    metadata: file_entry,
+                    content,
+                }))
+            }
+            Err(e) => Some(Err(Error::Io(e))),
         }
     }
 }